Refactor useQuery options to include dynamic page parameter
This commit refactors the useQuery options to include a dynamic page parameter. Previously, the currentPage variable was used to fetch movies from the API endpoint. Now, the page variable is utilized instead, providing more flexibility for dynamically changing the page number when fetching data. This change enhances the maintainability and scalability of the codebase.

Signed-off-by: Julia Stjerna <zerodeleo@gmail.com>