Remove watch function from movie list component
This commit removes the watch function from the movie list component. The watch function was previously used to monitor changes in the data and update related computed properties. However, it is no longer necessary after the refactoring of the component to include pagination functionality. The removal simplifies the code and improves its clarity.

Signed-off-by: Julia Stjerna <zerodeleo@gmail.com>