Fixed bug with infinite scroll
Resolved an issue with the infinite scroll feature where it was not functioning as expected. The bug caused the infinite scroll to not load additional movie data properly when the user reached the bottom of the page. This fix ensures that the infinite scroll feature works correctly, allowing users to load more movies seamlessly as they scroll.

Signed-off-by: Julia Stjerna <zerodeleo@gmail.com>